refactor repository wrappers out of data
